plat/network: Make Network component a driver for the Platform.
This commit restructure the network component such that it is a driver
for the platform component.  This prevents it from inheriting component
attributes which are otherwise reserved for repositories.
