+++ kraft/plat/network/driver/driver.py
@@ -1,0 +1,99 @@
+# SPDX-License-Identifier: BSD-3-Clause
+#
+# Authors: Alexander Jung <alexander.jung@neclab.eu>
+#
+# Copyright (c) 2020, NEC Europe Laboratories GmbH., NEC Corporation.
+#                     All rights reserved.
+#
+# Redistribution and use in source and binary forms, with or without
+# modification, are permitted provided that the following conditions
+# are met:
+#
+# 1. Redistributions of source code must retain the above copyright
+#    notice, this list of conditions and the following disclaimer.
+# 2. Redistributions in binary form must reproduce the above copyright
+#    notice, this list of conditions and the following disclaimer in the
+#    documentation and/or other materials provided with the distribution.
+# 3. Neither the name of the copyright holder nor the names of its
+#    contributors may be used to endorse or promote products derived from
+#    this software without specific prior written permission.
+#
+# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+# AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+# IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+# ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE
+# LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
+# CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
+# SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
+# INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
+# CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
+# ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+# POSSIBILITY OF SUCH DAMAGE.
+from __future__ import absolute_import
+from __future__ import unicode_literals
+
+from kraft.error import KraftError
+from kraft.error import NetworkDriverError
+
+
+class NetworkDriver(object):
+    _name = None
+    _type = None
+
+    @property
+    def name(self):
+        return self._name
+
+    @property
+    def type(self):
+        return self._type
+
+    def __init__(self, name=None, type=None):
+        if name is not None:
+            self._name = name
+        else:
+            self._name = NetworkDriver.generate_bridge_name()
+
+        self._type = type
+
+    def integrity_ok(self):
+        return False
+
+    def create_bridge(self, name=None, dry_run=False):
+        raise NetworkDriverError(
+            "Creating a bridge is not possible with driver %s" % self.type
+        )
+
+    def add_vif(self, name=None):
+        raise NetworkDriverError(
+            "Adding an interface is not possible with driver %s" % self.type
+        )
+
+    def remove_vif(self, name=None):
+        raise NetworkDriverError(
+            "Removing an interface is not possible with driver %s" % self.type
+        )
+
+    def destroy_bridge(self, name=None):
+        raise NetworkDriverError(
+            "Removing a bridge is not possible with driver %s" % self.type
+        )
+
+    def bridge_exists(self, name=None):
+        raise NetworkDriverError(
+            "Checking for a bridge is not possible with driver %s" % self.type
+        )
+
+    def generate_bridge_name(self, prefix='virbr', max_tries=1024):
+        suffix_i = 0
+        new_name = None
+
+        while suffix_i < max_tries:
+            new_name = prefix + str(suffix_i)
+
+            if not self.bridge_exists(new_name):
+                return new_name
+
+            suffix_i += 1
+
+        raise KraftError("Max tries for bridge creation reached!")
